+++ Classes/SchemaOrg/Language.php
@@ -3,7 +3,7 @@
 namespace Portrino\PxSemantic\SchemaOrg;
 
 /**
- * Natural languages such as Spanish, Tamil, Hindi, English, etc. and programming languages such as Scheme and Lisp.
+ * Natural languages such as Spanish, Tamil, Hindi, English, etc. Formal language code tags expressed in [BCP 47](https://en.wikipedia.org/wiki/IETF\_language\_tag) can be used via the [[alternateName]] property. The Language type previously also covered programming languages such as Scheme and Lisp, which are now best represented using [[ComputerLanguage]].
  *
  * @see http://schema.org/Language Documentation on Schema.org
  *